refactor: separate component folders in the stories directory
